Split System.cs into three files
